+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `StorageTransformerChain` now only accepts `Arc` storage.
  - Various group methods are now `#[must_use]`
  - Change `NodePath` internal representation to `PathBuf`
+ - Remove unneeded '/' prefix strip in `StorePrefix`
 
 ### Fixed
  - Fix store prefix to node path conversion and vice versa